[REFACTOR] lotto 패키지 정리
