+++ src/main/java/lotto/view/model/LottoResult.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.view.model;
 
 import java.util.Arrays;
 import java.util.HashMap;
