+++ src/test/java/lotto/LottoTicketTest.java
@@ -4,6 +4,9 @@ import lotto.exception.DuplicateNumberException;
 import lotto.exception.OutOfCountException;
 import lotto.exception.OutOfMaxNumberException;
 import common.NumberElementCollection;
+import lotto.view.domain.LottoTicket;
+import lotto.view.model.LottoResult;
+import lotto.view.model.LottoRule;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
